fix<ingredient>:chuyển qua Double không dùng bigdecimal cho số thập phân, và fix 1 số lỗi liên quan đêns validator,service

+++ SmartRestaurant/src/main/java/com/example/SmartRestaurant/dto/request/IngredientRequest.java
@@ -6,8 +6,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
@@ -16,6 +14,6 @@ public class IngredientRequest {
     private String name;
     private IngredientType type;
     private String unit;
-    private BigDecimal minStock;
-    private BigDecimal yieldRate;
+    private Double minStock;
+    private Double yieldRate;
 }
